pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task manager"
authors = ["Владимир Силин <sva@itx24.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
python-dotenv = "^1.0.1"
dj-database-url = "^2.3.0"
django-bootstrap5 = "^24.3"
gunicorn = "^23.0.0"
django-filter = "^24.3"
rollbar = "^1.1.2"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.9.0"
black = "^25.1.0"
ruff = "^0.9.3"
isort = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
